fix: refactor previews
